feat(CHAIN-3064): implement the main event loop for polling, proof generation, aggregation, and submission (#18)
* implement the main event loop for polling, proof generation, aggregation, and submission

* fix formatter warnings

* pr feedback

* improve tests

* improve tests

* fix clippy warnings

+++ crates/proof/proposer/crates/proposer/src/constants.rs
@@ -3,7 +3,7 @@
 use std::time::Duration;
 
 /// Number of proofs to aggregate in a single batch.
-pub const AGGREGATE_BATCH_SIZE: u64 = 512;
+pub const AGGREGATE_BATCH_SIZE: usize = 512;
 
 /// Number of blocks the EVM can look back for blockhashes.
 pub const BLOCKHASH_WINDOW: u64 = 256;
